updating indiv_models to use api-connector style definition
