+++ pytc/indiv_models/base.py
@@ -16,14 +16,38 @@ class ITCModel:
     Base class from which all ITC models should be sub-classed.
     """
 
-    # --------------------------------------------------------------------------
-    # These three methods need to be defined for each subclass.
+    param_definition = {}
 
-    def __init__(self):
-        pass
+    def __init__(self,
+                 S_cell=100e-6,S_syringe=0.0,
+                 T_cell=0.0,   T_syringe=1000e-6,
+                 cell_volume=300.0,
+                 shot_volumes=[2.5 for i in range(30)]):
 
-    def initialize_param(self):
-        pass
+        """
+        S_cell: stationary concentration in cell in M
+        S_syringe: stationary concentration in syringe in M
+        T_cell: titrant concentration cell in M
+        T_syringe: titrant concentration syringe in M
+        cell_volume: cell volume, in uL
+        shot_volumes: list of shot volumes, in uL.
+        shot_start: first shot to use in fit
+        """
+
+        self._S_cell = S_cell
+        self._S_syringe = S_syringe
+
+        self._T_cell = T_cell
+        self._T_syringe = T_syringe
+
+        self._cell_volume = cell_volume
+        self._shot_volumes = np.array(shot_volumes)
+
+        # Determine the concentration of all of the species across the titration
+        self._S_conc = self._titrate_species(self._S_cell,self._S_syringe)
+        self._T_conc = self._titrate_species(self._T_cell,self._T_syringe)
+
+        self._initialize_param()
 
     @property
     def dQ(self):
@@ -80,30 +104,16 @@ class ITCModel:
         Return the heat of dilution.
         """
 
-        #return self.mole_ratio*self.param_values["dilution_heat"] + self.param_values["dilution_intercept"]
         return self._T_conc[1:]*self.param_values["dilution_heat"] + self.param_values["dilution_intercept"]
 
-    def _initialize_param(self,param_names=None,param_guesses=None):
+    def _initialize_param(self):
         """
-        Initialize the parameters.  If param_names is None, the parameter
-        names and guesses are determined by inspection of self.initialize_param.
+        Initialize the parameters.  
         """
 
-        self._params = {}
-        if param_names == None:
-
-            # If parameter names are not specified, grab them from
-            # self.initialize_param defaults
-            a = inspect.getargspec(self.initialize_param)
-            param_names = a.args
-            param_names.remove("self")
-            param_guesses = a.defaults
-
-        for i, p in enumerate(param_names):
-            if param_guesses != None:
-                self._params[p] = fit_param.FitParameter(p,guess=param_guesses[i])
-            else:
-                self._params[p] = fit_param.FitParameter(p)
+        self._params = {"dilution_heat":0.0,"dilution_interecept":0.0}
+        for p in self.param_definition.keys():
+            self._params[p] = fit_param.FitParameter(p,guess=self.param_definition[p])
 
         self._param_names = param_names[:]
         self._param_names.sort()
